Officially support building assembler files
erlc is wired to treat *.S files as assembler and build them as
compile:file(File, [from_asm]), but this is not documented. There's also
a documented compile:file/2 option called 'asm' (mapping to 'from_asm'),
but the wording discourages its use. All of this has been in place and
in use for a long time. Therefore, it should be supported officially.

To fix that, make the following changes:
 * document erlc handling of *.core files
 * un-document 'asm' and document 'from_asm' instead
 * deprecate 'asm'

While at it, fix a minor typo in the test suite.