Officially support building core files
erlc is wired to treat *.core files as core and build them as
compile:file(File, [from_core]), but this is not documented. There's
also an udocumented compile:file/2 option called 'from_core'. This has
been in place and in use for a long time. Therefore, it should be
supported officially.

To fix that, make the following changes:
 * document erlc handling of *.core files
 * document 'from_core'