Generalize optimizations of case statements
Case expressions such as:

  case {Expr1,Expr} of
     {V1,V2} -> ...
  end

are already optimized to not actually build the tuple. Generalize
the optimization to avoid building any kind of composite term,
such as:

  case {ok,[A,B]} of
    {ok,[X,Y]} -> ...
  end

We don't expect programmers to write such code directly, but
inlining can produce such code.

We need to be careful about the warnings we produce. If the case
expression is a literal, it is expected that no warnings should be
produced for clauses that don't match. We must make sure that we
continue to suppress those warnings.