Suppress error report when child was terminated normally
Description:
Let's assume we have a supervisor_bridge as a
non-permanent child of a supervisor. When the
bridge terminates normally it reports by
error_logger:error_report, but the supervisor does
nothing. So we have tons
(especially, when it's simple_one_for_one) of
error messages while everything is completely ok.

Let's assume we have a supervisor_bridge as a
permanent child of a supervisor. When the bridge
terminates, it invokes error_logger, but what for?
The supervisor will invoke error_logger about this
error right after that.
So what is the reason for the error_logger to log
for normally terminating children in this case?