Don't make gzio.c dependent on the zutil.h header file
gzio.c is our own replacement for zlib's gzopen() etc
(based on a version of gzio.c that was included in an old version
of zlib).

Unfortunately, gzio.c still depends on the *internal* zlib header file
zutil.h which is not supposed to be used outside of the zlib source
code. The dependencies are the use of the gzFile typedef and the
F_OPEN() macro.

Instead of gzFile, define and use our own ErtsGzFile.

To get rid of the F_OPEN() macro, call open() of _wfopen() directly.