configure: Prefer the system's zlib over own our zlib source
WxWidgets (used by the wx application) also uses zlib. To
ensure that the run-time system and WxWidgets use the same version
of zlib, use the system's zlib if present.

Also, the system's zlib may be specially optimized and thus faster
than our own generic source code.

We only use zlib versions that are "good enough". For now, that
means 1.2.5 or higher.

Remove the option --enable-shared-zlib and add the option
--enable-builtin-zlib to force the use of the built-in zlib.