wx: Fix hanging wx calls
wxWakeUpIdle doesn't always work on wxGTK and the workaround is to
periodically invoke it, so the gui-thread doesn't get stuck in poll.

Previously it was only called when NOT in batch mode, however if the
wxWakeUpIdle call after a batch_begin command did not wake up the
thread it would still get stuck in poll.