Add application_controller:change_config_data/1
Currently the application_controller loads configuration files on
bootstrap by directly checking the values given to the -config flag.
The configuration data is kept by the application controller and
merged into the application environment on load. There is no mechanism
to change the configuration data besides during boot.

This commit adds a new application_controller:change_config_data/1
function that allows the configuration data to be changed, useful
when they need to be read from other sources besides the ones given
to -config at boot. The new function only changes the configuration
data, applications that are already loaded are not affected by this
function.