Allow making the system with --enable-native-libs
The HiPE compiler crashes when trying to compile these files because
it does not currently support maps. So, add a -compile(no_native)
attribute to these files to allow the system to be made even when
configured with --enable-native-libs.

This is a temporary fix and will be removed when the HiPE compiler
gets proper support for maps.