asn1ct_table: Remove unused flexibility in table creation
Unused flexibility complicates the code and can hide bugs.

new/1 is never called with a table name that already exists, so
we should remove the code that allows for that. That will increase
the coverage of asn1ct_table to 100%.

Also remove the new/2 and new_reuse/2 functions which are never
called externally.