compiler: Fix codegen multiple updates for Maps
This fixes an error on multiple updates optimization for map pairs.
The error was introduced with moving to term order in Maps.

This also fixes an error where register life time was lost for values
and could result in erroneuos values being emitted in for map pairs.

Simplified v3_codegen by moving multiple update optimizations to v3_kernel.