Properly silence an unmatched return warning
An underscore, presumably used to signify that a file:delete/1 call
was to be issued independently of whether the call would work ok or
not, was placed at a point that made dialyzer -Wunmatched_returns
on this file complain that the result of the case expression was
an unmatched union of an 'ok' and and {'error',term()} type value.
Place the underscore in its proper place instead.