otp_build: Include parsetools in the release of a small system
When releasing a small system like this:

  ./otp_build setup
  ./otp_build release TARGET_DIR

include the parsetools application.

In one of our internal scripts, a previously built and released
Erlang/OTP system is used to update the primary bootstrap in a
git repository. A small system could not be used for that purpose
because parsetools was missing.