beam/utils.c: Change back cmp() to only take two arguments
Commit d5c238473b9cec819d93faaef4ccc00ddb60465f added a third
argument to the cmp() function, but did not change the code generation
of native code that called cmp().

Name the new functions with an "erts_" prefix, and reinstate the
old cmp() function with two arguments.

We don't have to keep a cmp() function in the halfword emulator,
since HiPE does not support the halfword emulator.