ose: Thread priorities configurable from lmconf
The pattern used for getting the priority from the lmconf
is based on the name of the process created. The pattern is:

ERTS_%%PROCESS_NAME%%_PRIO

with the %%PROCESS_NAME%% replaced by the prefix of the process
the priority applies to. eg:

ERTS_SCHEDULER_PRIO=24

applies to processes with name SCHEDULER_1, SCHEDULER_2 etc.