Change spawn/1 + monitor/2 to spawn_monitor/1 to avoid deadlock
crashdump_viewer:progress_pmap did sometimes hang since is spawned a
process, then monitored it and waited for a specific DOWN
message. When the process' work was very fast, it would exit before
the call to monitor and the DOWN message would contain reason 'noproc'
instead of the expected {pmap_done,Result}.

By doing spawn_monitor/1 instead, the monitor is always set before the
process exits so the deadlock is avoided.