epp: Make it possible to specify a default encoding
In the next commit, we will need a way to tell epp which the
default encoding should be for files that have no encoding comment.

We could add new open() and parse_file() functions with one
extra argument for the encoding, but there are already too many
variants.

To avoid having to add an additional argument to epp:open() and
epp:parse_file() each time new options are needed, introduce
epp:open/1 and epp:parse_file/2 that takes a property list with
options. Also support the new 'default_encoding' option for specifying
the default encoding for source files.

Thanks to Richard Carlsson for the idea and the implementation
of the new functionality in epp.erl.