Introduce compiler option 'nowarn_deprecated_type'
The deprecation of the built-in types dict/0 and so on had as
side-effect that it was impossible to switch to dict:dict/2 and so on
without getting warnings either in the the previous release (R16B) or
the current one (17.0).

By including the attribute

    -compile(nowarn_deprecated_type).

in an Erlang source file warnings about deprecated types can be
avoided in 17.0.

The option can also be given as a compiler flag:

erlc +nowarn_deprecated_type file.erl