stdlib: Fix a Dets bug where files were deleted
Instead of deleting the small file when opening it with
dets:open_file/2 an error message is returned, mimicking the behaviour
when the file is bigger.

open_file/1 has been modified to return the same error message when
the file is small as when the file is bigger.

Thanks to Hakan Mattson for reporting this bug.