Fix crash in crashdump_viewer when node has multiple creations
A node to which we have references to multiple instances (creations)
will have an information line in the crashdump like this:

Creation: 1 2 ...

This would earlier crash because crashdump_viewer would try to do
list_to_integer on the value after "Creation: ". This is now
corrected.

This correction also helps the case when the emulator is debug
compiled, since the line could then be

Creation: 1 (refc=1)