sasl: Fix crash in systools_make:format_error/1
The crash occurred when systools:make_script/2 raises the
'duplicate_modules' error. Depending on the 'silent' option,
systools_make:format_error/1 is called to either format and display an
error message, or return a tuple containing the error properties.

When displaying a 'duplicate_modules' error message, format_error/1
receives a list of applications providing the same module(s):
  [
    {{Mod,App1,_}, {Mod,App2,_}},
    ...
  ]

However, before this fix, format_error/1 expected the following
structure, leading to a 'function_clause' exception:
  [
    {{Mod,_,App1,_,_}, {Mod,_,App2,_,_}},
    ...
  ]

The crash never occurred with the 'silent' option, because the error and
its properties are returned as is to the caller.