Extract line annotation from filter qualifiers in a type-friendly way
A recently submitted patch used function get_anno/1 to extract the line
annotation from filter qualifiers in comprehensions, but this did not
respect the spec of this function and resulted in a dialyzer warning.
To make the code more type-friendly, an appropriate get_qual_line/1
function was introduced.  The new function also ensures that the
qualifiers are of the appropriate form.

While at it, also simplified an Edoc comment.