stdlib: Generalize erl_parse:abstract/2
The 'encoding' option of erl_parse:abstract/2 has been extended to
include 'none' and a callback function (a predicate).

The rationale is that a more general means of determining what integer
lists are to be represented as strings may help readability when
generating Erlang code given input in some other encoding than Latin-1
or UTF-8.