Shorten the Introduction section, keeping only the essential details
We can assume that anyone that reads the documentation for the Asn1
documentation already knows about ASN.1, so we don't need three
paragraphs of introductory. Keep one short paragraph explaining
what ASN.1 is in case a reader unfamiliar with ASN.1 stumbles upon
the manual.

While we are at it, reformat the paragraphs in Introduction to
shorter lines that don't wrap.